+++ packages/bpk-component-nudger/index.js
@@ -1,0 +1,3 @@
+import BpkNudger from './src/BpkNudger';
+
+export default BpkNudger;