+++ changelog.md
@@ -5,6 +5,8 @@
 **Added:**
 - bpk-component-accordion:
   - New accordion component (http://backpack.prod.aws.skyscnr.com/docs/molecules/accordions/)
+- bpk-component-nudger:
+  - New nudger component (http://backpack.prod.aws.skyscnr.com/docs/molecules/nudgers/)
 - bpk-component-checkbox:
   - Added ability to add custom className and other arbitrary props
 - bpk-component-paragraph: